// frontend/src/app/models/konobar.ts
export class Konobar{
    ime: string="";
    prezime: string="";
    pol:string="";
    kor_ime: string="";
    lozinka: string="";
    bezb_pitanje="";
    bezb_odgovor="";
    telefon: number=0;
    adresa:string="";
    mejl: string="";
    img:string="";
    restoran:string="";
    status:string=""

}